trying to refactor the getEndpoint to get the trackingNumbers and run the tracking requests, which will then be returned to the UI
