+++ package.json
@@ -7,6 +7,7 @@
     "start": "node app.js"
   },
   "dependencies": {
+    "async": "^2.1.4",
     "body-parser": "^1.15.2",
     "cfenv": "^1.0.3",
     "express": "^4.14.0",
